:truck: :lipstick: Moved scripts to the Editor Folder. Changed mesh name
