+++ unity/Holo/Assets/Editor/STLConversion/STLImportUtils.cs
